+++ lib/writter.ex
@@ -1,7 +1,5 @@
 defmodule Xcribe.Writter do
-  @moduledoc """
-  Writes the documentation to an file on the project.
-  """
+  @moduledoc false
 
   alias Xcribe.Config
 
@@ -18,6 +16,7 @@ defmodule Xcribe.Writter do
     {:ok, file} = File.open(output_file, [:write])
 
     IO.binwrite(file, text)
+    IO.puts("#{IO.ANSI.cyan()}> Xcribe documentation written in #{output_file}#{IO.ANSI.reset()}")
 
     File.close(file)
   end