+++ test/support/samples/swagger_formater/path_ithem_object.ex
@@ -1,5 +1,5 @@
-defmodule Xcribe.Support.Samples.SwaggerFormater do
-  def path_item_object_without_request_body do
+defmodule Xcribe.Support.Samples.SwaggerFormater.PathItemObject do
+  def not_found_without_req_body do
     %{
       "put" => %{
         summary: "",
@@ -32,7 +32,7 @@ defmodule Xcribe.Support.Samples.SwaggerFormater do
     }
   end
 
-  def path_item_object_with_request_body do
+  def success_with_req_body do
     %{
       "put" => %{
         summary: "",
@@ -97,7 +97,7 @@ defmodule Xcribe.Support.Samples.SwaggerFormater do
     }
   end
 
-  def expected_path_objects_merge do
+  def bad_request_with_req_body do
     %{
       "put" => %{
         summary: "",
@@ -105,17 +105,77 @@ defmodule Xcribe.Support.Samples.SwaggerFormater do
         security: [%{"bearer" => []}],
         parameters: [
           %{
+            example: "3456efae-bc48-44b4-94d0-7670c4772696",
+            in: "path",
+            name: "id",
+            required: true,
+            schema: %{type: "string"}
+          },
+          %{
             name: "include",
             in: "query",
             schema: %{type: "string"},
             example: "user"
+          }
+        ],
+        responses: %{
+          400 => %{
+            description: "",
+            content: %{
+              "application/json" => %{schema: %{type: "string", example: "bad request"}}
+            },
+            headers: %{
+              "cache-control" => %{description: "", schema: %{type: "string"}},
+              "x-request-id" => %{description: "", schema: %{type: "string"}}
+            }
+          }
+        },
+        requestBody: %{
+          description: "",
+          content: %{
+            "multipart/mixed" => %{
+              schema: %{
+                type: "object",
+                properties: %{
+                  "address" => %{
+                    properties: %{
+                      "city" => %{type: "string"}
+                    },
+                    type: "object"
+                  },
+                  "cnpj" => %{type: "string"},
+                  "company_name" => %{type: "string"},
+                  "email" => %{type: "string"},
+                  "phone" => %{type: "string"},
+                  "trading_name" => %{type: "string"}
+                }
+              }
+            }
+          }
+        }
+      }
+    }
+  end
+
+  def all_merged do
+    %{
+      "put" => %{
+        summary: "",
+        description: "",
+        security: [%{"bearer" => []}],
+        parameters: [
+          %{
+            example: "3456efae-bc48-44b4-94d0-7670c4772696",
+            in: "path",
+            name: "id",
+            required: true,
+            schema: %{type: "string"}
           },
           %{
-            name: "id",
-            in: "path",
-            required: true,
+            name: "include",
+            in: "query",
             schema: %{type: "string"},
-            example: "590eda29-df5f-4244-8159-7bb87e3cc15a"
+            example: "user"
           }
         ],
         responses: %{
@@ -123,6 +183,16 @@ defmodule Xcribe.Support.Samples.SwaggerFormater do
             description: "",
             content: %{
               "application/json" => %{schema: %{type: "string", example: "success"}}
+            },
+            headers: %{
+              "cache-control" => %{description: "", schema: %{type: "string"}},
+              "x-request-id" => %{description: "", schema: %{type: "string"}}
+            }
+          },
+          400 => %{
+            description: "",
+            content: %{
+              "application/json" => %{schema: %{type: "string", example: "bad request"}}
             },
             headers: %{
               "cache-control" => %{description: "", schema: %{type: "string"}},